Correção Entidades

+++ root/src/main/java/br/com/cesarschool/poo/titulos/entidades/Acao.java
@@ -16,7 +16,7 @@ import java.time.LocalDate;
 public class Acao extends Ativo{
 private double valorUnitario;
 
-public Acao(int identificador, String nome, LocalDate dataValidade, double valorUnitario){
+public Acao(long identificador, String nome, LocalDate dataValidade, double valorUnitario){
 super(identificador, nome, dataValidade);
 this.valorUnitario=valorUnitario;
 }
