+++ root/src/main/java/br/com/cesarschool/poo/titulos/entidades/TituloDivida.java
@@ -17,7 +17,7 @@ import java.time.LocalDate;
 public class TituloDivida extends Ativo{
     private double taxaJuros;
 
-    public TituloDivida(int identificador, String nome, LocalDate dataValidade, double taxaJuros){
+    public TituloDivida(long identificador, String nome, LocalDate dataValidade, double taxaJuros){
         super(identificador, nome, dataValidade);
         this.taxaJuros=taxaJuros;
     }
@@ -25,7 +25,6 @@ public class TituloDivida extends Ativo{
     public double getTaxaJuros() {
         return taxaJuros;
     }
-
     public void setTaxaJuros(double taxaJuros) {
         this.taxaJuros = taxaJuros;
     }
